fix: keep value signal for abs min max
The absolute values for min and max were meant to keep the the sign of
the value. The abs_min and abs_max should only do an absolute
comparison, not change the value. Change stored value to keep the sign
of the value, and not the absolute value of it.